feat: closures, upvalues, closing upvalues

+++ include/debug.h
@@ -4,6 +4,6 @@
 #include "chunk.h"
 
 void disassemble_chunk(chunk_t *chunk, const char *name);
-size_t disassemble_instruction(chunk_t *chunk, size_t offset);
+int disassemble_instruction(chunk_t *chunk, int offset);
 
 #endif // GUSTAV_DEBUG_H
